make a custom scrubber for internalized exceptions
