+++ tests/Speckle.Sdk.Serialization.Tests/ExceptionTests.Test_Exceptions_Receive_Server.verified.json
@@ -1,6 +1,9 @@
 ﻿{
-  "Data": {},
   "Message": "The method or operation is not implemented.",
-  "StackTrace": "at Speckle.Sdk.Serialization.Tests.ExceptionServerObjectManager.DownloadSingleObject(String objectId, IProgress`1 progress, CancellationToken cancellationToken)\nat Speckle.Sdk.Serialisation.V2.Receive.ObjectLoader.GetAndCache(String rootId, DeserializeProcessOptions options, CancellationToken cancellationToken)\nat Speckle.Sdk.Serialisation.V2.Receive.DeserializeProcess.Deserialize(String rootId)\n--- End of stack trace from previous location ---\nat Xunit.Assert.RecordExceptionAsync(Func`1 testCode)",
-  "Type": "NotImplementedException"
+  "Data": {},
+  "InnerException": null,
+  "HelpLink": null,
+  "Source": "Speckle.Sdk.Serialization.Tests",
+  "HResult": -2147467263,
+  "StackTrace": "   at Speckle.Sdk.Serialization.Tests.ExceptionServerObjectManager.DownloadSingleObject(String objectId, IProgress`1 progress, CancellationToken cancellationToken) in C:\\Users\\adam\\Git\\speckle-sharp-sdk\\tests\\Speckle.Sdk.Serialization.Tests\\Framework\\ExceptionServerObjectManager.cs:line 19\r\n   at Speckle.Sdk.Serialisation.V2.Receive.ObjectLoader.GetAndCache(String rootId, DeserializeProcessOptions options, CancellationToken cancellationToken) in C:\\Users\\adam\\Git\\speckle-sharp-sdk\\src\\Speckle.Sdk\\Serialisation\\V2\\Receive\\ObjectLoader.cs:line 52\r\n   at Speckle.Sdk.Serialisation.V2.Receive.DeserializeProcess.Deserialize(String rootId) in C:\\Users\\adam\\Git\\speckle-sharp-sdk\\src\\Speckle.Sdk\\Serialisation\\V2\\Receive\\DeserializeProcess.cs:line 42\r\n   at Speckle.Sdk.Serialization.Tests.ExceptionTests.<>c__DisplayClass3_0.<<Test_Exceptions_Receive_Server>b__0>d.MoveNext() in C:\\Users\\adam\\Git\\speckle-sharp-sdk\\tests\\Speckle.Sdk.Serialization.Tests\\ExceptionTests.cs:line 74\r\n--- End of stack trace from previous location ---\r\n   at Xunit.Assert.RecordExceptionAsync(Func`1 testCode) in /_/src/xunit.assert/Asserts/Record.cs:line 114"
 }
