+++ tests/Speckle.Sdk.Serialization.Tests/Framework/TestFileManager.cs
